+++ app/src/main/java/com/quigglesproductions/secureimageviewer/models/modular/file/ModularFile.java
@@ -165,8 +165,8 @@ public class ModularFile implements ItemBaseModel, IDisplayFile {
     }
 
     @Override
-    public long getId() {
-        return onlineId;
+    public Long getId() {
+        return (long) onlineId;
     }
 
     public IFileDataSource getDataSource() {
